requirements.txt
streamlit
Pillow
plotly
networkx
pyvis
gspread
oauth2client
